Move ConnectionPool and ConnectionDestination classes from API into Client module.
