+++ components/client/src/main/java/com/hotels/styx/client/connectionpool/ConnectionPool.java
@@ -13,8 +13,10 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-package com.hotels.styx.api.client;
+package com.hotels.styx.client.connectionpool;
 
+import com.hotels.styx.api.client.Connection;
+import com.hotels.styx.api.client.Origin;
 import rx.Observable;
 
 import java.io.Closeable;
