+++ components/client/src/main/java/com/hotels/styx/client/connectionpool/ConnectionPoolFactory.java
@@ -16,7 +16,6 @@
 package com.hotels.styx.client.connectionpool;
 
 import com.hotels.styx.api.client.Connection;
-import com.hotels.styx.api.client.ConnectionPool;
 import com.hotels.styx.api.client.Origin;
 import com.hotels.styx.api.metrics.MetricRegistry;
 import com.hotels.styx.api.service.ConnectionPoolSettings;
