+++ components/client/src/test/unit/java/com/hotels/styx/client/StyxHostHttpClientTest.java
@@ -16,7 +16,7 @@
 package com.hotels.styx.client;
 
 import com.hotels.styx.api.Id;
-import com.hotels.styx.api.client.ConnectionPool;
+import com.hotels.styx.client.connectionpool.ConnectionPool;
 import com.hotels.styx.api.FullHttpRequest;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
