+++ components/client/src/test/unit/java/com/hotels/styx/client/TransportTest.java
@@ -21,7 +21,7 @@ import com.hotels.styx.api.HttpResponse;
 import com.hotels.styx.api.Id;
 import com.hotels.styx.api.StyxObservable;
 import com.hotels.styx.api.client.Connection;
-import com.hotels.styx.api.client.ConnectionPool;
+import com.hotels.styx.client.connectionpool.ConnectionPool;
 import com.hotels.styx.api.exceptions.NoAvailableHostsException;
 import io.netty.buffer.ByteBuf;
 import org.testng.annotations.BeforeMethod;
