+++ components/client/src/test/unit/java/com/hotels/styx/client/netty/connectionpool/StubConnectionPool.java
@@ -17,7 +17,7 @@ package com.hotels.styx.client.netty.connectionpool;
 
 import com.google.common.base.Objects;
 import com.hotels.styx.api.client.Connection;
-import com.hotels.styx.api.client.ConnectionPool;
+import com.hotels.styx.client.connectionpool.ConnectionPool;
 import com.hotels.styx.api.service.ConnectionPoolSettings;
 import com.hotels.styx.api.client.Origin;
 import com.hotels.styx.client.connectionpool.stubs.StubConnectionFactory;
