+++ components/proxy/src/test/java/com/hotels/styx/admin/dashboard/DashboardDataTest.java
@@ -19,7 +19,7 @@ import com.codahale.metrics.Gauge;
 import com.google.common.eventbus.EventBus;
 import com.hotels.styx.Version;
 import com.hotels.styx.api.HttpHandler;
-import com.hotels.styx.api.client.ConnectionPool;
+import com.hotels.styx.client.connectionpool.ConnectionPool;
 import com.hotels.styx.api.client.Origin;
 import com.hotels.styx.api.client.OriginsSnapshot;
 import com.hotels.styx.api.client.RemoteHost;
