+++ components/proxy/src/test/java/com/hotels/styx/admin/tasks/StubConnectionPool.java
@@ -16,7 +16,7 @@
 package com.hotels.styx.admin.tasks;
 
 import com.hotels.styx.api.client.Connection;
-import com.hotels.styx.api.client.ConnectionPool;
+import com.hotels.styx.client.connectionpool.ConnectionPool;
 import com.hotels.styx.api.client.Origin;
 import com.hotels.styx.api.service.ConnectionPoolSettings;
 import rx.Observable;
